added authorization

+++ backend/routes/posts.js
@@ -38,7 +38,8 @@ router.post('', checkAuth,
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then(createdPost => {
     res.status(201).json({
@@ -67,12 +68,16 @@ router.put('/:id',
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator: req.userData.userId
     });
-    console.log(post);
-    Post.updateOne({_id: req.params.id}, post)
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId }, post)
       .then(result => {
-        res.status(200).json({message: 'Update successful!'});
+        if (result.nModified > 0) {
+          res.status(200).json({message: 'Update successful!'});
+        } else {
+          res.status(401).json({message: 'Not Authorized!'});
+        }
       });
 });
 
@@ -111,11 +116,15 @@ router.get('/:id', (req, res, next) => {
     })
 });
 
-router.delete('/:id', (req, res, next) => {
-  Post.deleteOne({_id: req.params.id})
+router.delete('/:id', checkAuth, (req, res, next) => {
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId })
     .then(result => {
       console.log(result);
-      res.status(200).json({message: 'Post deleted!'});
+      if (result.n > 0) {
+        res.status(200).json({message: 'Deletion successful!'});
+      } else {
+        res.status(401).json({message: 'Not Authorized!'});
+      }
     });
 });
